changes per code-review + bugfixes
