+++ modules/mcp/abilities/manage-classes-ability.php
@@ -16,6 +16,7 @@ use Elementor\Modules\GlobalClasses\Global_Classes_Labels;
 use Elementor\Modules\GlobalClasses\Global_Classes_Repository;
 use Elementor\Modules\GlobalClasses\Global_Classes_REST_API;
 use Elementor\Modules\Mcp\Abilities\Utils\Bulk_Operations_Result;
+use Elementor\Modules\Mcp\Abilities\Utils\Style_Variants_Merger;
 use Elementor\Modules\Variables\Module as Variables_Module;
 use Elementor\Modules\Variables\Services\Batch_Operations\Batch_Processor;
 use Elementor\Modules\Variables\Services\Variables_Service;
@@ -31,7 +32,6 @@ class Manage_Classes_Ability extends Abstract_Ability {
 	const CLASS_TYPE = 'class';
 	const DESKTOP_BREAKPOINT = 'desktop';
 	const MAX_BATCH_SIZE = 50;
-	const PSEUDO_STATES = [ 'hover', 'focus', 'active' ];
 
 	private ?Global_Classes_Repository $repository;
 	private ?Css_Converter $css_converter;
@@ -89,7 +89,7 @@ class Manage_Classes_Ability extends Abstract_Ability {
 								'label' => [ 'type' => 'string' ],
 								'css' => [
 									'type' => 'object',
-									'description' => 'Legacy: flat property → value map. Ignored when styles is present.',
+									'description' => 'Flat property → value map for CSS declarations. Ignored when styles is present.',
 								],
 								'styles' => [
 									'type' => 'object',
@@ -261,7 +261,7 @@ class Manage_Classes_Ability extends Abstract_Ability {
 				'id'                  => $class_id,
 				'label'               => $label,
 				'css'                 => [],
-				'styles_tuples'       => $parsed['styles_tuples'],
+				'breakpoint_blocks'   => $parsed['breakpoint_blocks'],
 				'removal_breakpoints' => $parsed['removal_breakpoints'],
 			];
 			$intents['added_ids'][] = $class_id;
@@ -295,6 +295,12 @@ class Manage_Classes_Ability extends Abstract_Ability {
 			return;
 		}
 
+		$mode = $operation['mode'] ?? 'patch';
+		if ( ! in_array( $mode, [ 'patch', 'replace' ], true ) ) {
+			$results->add_error( $index, 'update', 'invalid_input', sprintf( 'Unknown mode: %s. Valid modes: patch, replace.', $mode ) );
+			return;
+		}
+
 		if ( ! in_array( $id, $current_order, true ) || isset( $deleted_set[ $id ] ) ) {
 			$results->add_error( $index, 'update', 'class_not_found', __( 'Global class not found', 'elementor' ) );
 			return;
@@ -316,7 +322,7 @@ class Manage_Classes_Ability extends Abstract_Ability {
 				'id'                  => $id,
 				'label'               => $label,
 				'css'                 => [],
-				'styles_tuples'       => $parsed['styles_tuples'],
+				'breakpoint_blocks'   => $parsed['breakpoint_blocks'],
 				'removal_breakpoints' => $parsed['removal_breakpoints'],
 				'mode'                => $operation['mode'] ?? 'patch',
 				'existing_variants'   => $existing['variants'] ?? [],
@@ -341,7 +347,7 @@ class Manage_Classes_Ability extends Abstract_Ability {
 			'css'              => $css,
 			'mode'             => $operation['mode'] ?? 'patch',
 			'existing_variants' => $existing['variants'] ?? [],
-			'null_reset_props' => $this->extract_null_reset_props( $css ),
+			'null_reset_props' => Style_Variants_Merger::extract_null_reset_props( $css ),
 		];
 
 		if ( ! in_array( $id, $intents['modified_ids'], true ) ) {
@@ -355,13 +361,13 @@ class Manage_Classes_Ability extends Abstract_Ability {
 
 	private function parse_styles_field( int $index, string $action, array $operation, Bulk_Operations_Result $results ): ?array {
 		$raw_styles          = $operation['styles'];
-		$styles_tuples       = [];
+		$breakpoint_blocks   = [];
 		$removal_breakpoints = [];
+		$active_breakpoints  = $this->get_active_breakpoint_keys();
 
 		foreach ( $raw_styles as $key => $css_string ) {
 			$breakpoint = 'default' === $key ? self::DESKTOP_BREAKPOINT : $key;
 
-			$active_breakpoints = $this->get_active_breakpoint_keys();
 			if ( ! in_array( $breakpoint, $active_breakpoints, true ) ) {
 				$results->add_error(
 					$index,
@@ -384,14 +390,14 @@ class Manage_Classes_Ability extends Abstract_Ability {
 				return null;
 			}
 
-			$styles_tuples[] = [
+			$breakpoint_blocks[] = [
 				'breakpoint' => $breakpoint,
 				'blocks'     => $result['blocks'],
 			];
 		}
 
 		return [
-			'styles_tuples'       => $styles_tuples,
+			'breakpoint_blocks'   => $breakpoint_blocks,
 			'removal_breakpoints' => $removal_breakpoints,
 		];
 	}
@@ -434,8 +440,8 @@ class Manage_Classes_Ability extends Abstract_Ability {
 		$touched_items = [];
 
 		foreach ( $intents['creates'] as $index => $intent ) {
-			if ( isset( $intent['styles_tuples'] ) ) {
-				$variants                      = $this->build_variants_from_styles_tuples( $intent['styles_tuples'] );
+			if ( isset( $intent['breakpoint_blocks'] ) ) {
+				$variants                      = Style_Variants_Merger::build_variants( $intent['breakpoint_blocks'], $this->get_css_converter() );
 				$touched_items[ $intent['id'] ] = [
 					'id'       => $intent['id'],
 					'label'    => $intent['label'],
@@ -456,8 +462,8 @@ class Manage_Classes_Ability extends Abstract_Ability {
 		}
 
 		foreach ( $intents['updates'] as $index => $intent ) {
-			if ( isset( $intent['styles_tuples'] ) ) {
-				$new_variants            = $this->build_variants_from_styles_tuples( $intent['styles_tuples'] );
+			if ( isset( $intent['breakpoint_blocks'] ) ) {
+				$new_variants            = Style_Variants_Merger::build_variants( $intent['breakpoint_blocks'], $this->get_css_converter() );
 				$removal_breakpoints     = $intent['removal_breakpoints'];
 				$existing_after_removal  = array_values(
 					array_filter(
@@ -465,11 +471,11 @@ class Manage_Classes_Ability extends Abstract_Ability {
 						fn( $v ) => ! in_array( $v['meta']['breakpoint'] ?? null, $removal_breakpoints, true )
 					)
 				);
-				$merged_variants = $this->apply_mode(
+				$merged_variants = Style_Variants_Merger::apply_mode(
 					$existing_after_removal,
 					$new_variants,
 					$intent['mode'],
-					array_column( $intent['styles_tuples'], 'breakpoint' )
+					array_column( $intent['breakpoint_blocks'], 'breakpoint' )
 				);
 				$touched_items[ $intent['id'] ] = [
 					'id'       => $intent['id'],
@@ -487,7 +493,7 @@ class Manage_Classes_Ability extends Abstract_Ability {
 				continue;
 			}
 
-			$class_item['variants'] = $this->apply_mode(
+			$class_item['variants'] = Style_Variants_Merger::apply_mode(
 				$intent['existing_variants'],
 				$class_item['variants'],
 				$intent['mode'],
@@ -499,101 +505,6 @@ class Manage_Classes_Ability extends Abstract_Ability {
 		}
 
 		return $touched_items;
-	}
-
-	private function extract_null_props( string $css ): array {
-		$null_props = [];
-		preg_replace_callback(
-			'/([a-zA-Z][a-zA-Z0-9-]*)\s*:\s*null\s*;?/',
-			function ( $matches ) use ( &$null_props ) {
-				$null_props[] = $matches[1];
-			},
-			$css
-		);
-		return $null_props;
-	}
-
-	private function strip_null_declarations( string $css ): string {
-		return (string) preg_replace( '/([a-zA-Z][a-zA-Z0-9-]*)\s*:\s*null\s*;?/', '', $css );
-	}
-
-	private function build_variants_from_styles_tuples( array $styles_tuples ): array {
-		$variants = [];
-
-		foreach ( $styles_tuples as $tuple ) {
-			$bp                    = $tuple['breakpoint'];
-			$blocks                = $tuple['blocks'];
-			$base_block_css        = '';
-			$base_custom_css_parts = [];
-
-			foreach ( $blocks as $block ) {
-				$selector = $block['selector'];
-				$css      = $block['css'];
-
-				if ( null === $selector ) {
-					$base_block_css = $css;
-					continue;
-				}
-
-				$state = ltrim( $selector, ':' );
-
-				if ( in_array( $state, self::PSEUDO_STATES, true ) ) {
-					$null_props     = $this->extract_null_props( $css );
-					$stripped_css   = $this->strip_null_declarations( $css );
-					$result         = $this->get_css_converter()->convert( $stripped_css );
-					$props          = $result['props'] ?? [];
-					$custom_css_str = $result['customCss'] ?? '';
-
-					if ( empty( $props ) && '' === $custom_css_str && empty( $null_props ) ) {
-						continue;
-					}
-
-					$variants[] = [
-						'meta'       => [
-							'breakpoint' => $bp,
-							'state'      => $state,
-						],
-						'props'      => $props,
-						'custom_css' => '' !== $custom_css_str
-							? [ 'raw' => base64_encode( $custom_css_str ) ] // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_encode -- Global class custom_css.raw is stored as base64.
-							: null,
-						'null_props' => $null_props,
-					];
-					continue;
-				}
-
-				$base_custom_css_parts[] = '&' . $selector . ' { ' . trim( $css ) . ' }';
-			}
-
-			$base_null_props = $this->extract_null_props( $base_block_css );
-			$base_block_css  = $this->strip_null_declarations( $base_block_css );
-			$base_result     = $this->get_css_converter()->convert( $base_block_css );
-			$base_props      = $base_result['props'] ?? [];
-			$base_custom_str = $base_result['customCss'] ?? '';
-
-			if ( ! empty( $base_custom_css_parts ) ) {
-				$extra           = implode( ' ', $base_custom_css_parts );
-				$base_custom_str = '' !== $base_custom_str ? $base_custom_str . ' ' . $extra : $extra;
-			}
-
-			if ( empty( $base_props ) && '' === $base_custom_str && empty( $base_null_props ) ) {
-				continue;
-			}
-
-			$variants[] = [
-				'meta'       => [
-					'breakpoint' => $bp,
-					'state'      => null,
-				],
-				'props'      => $base_props,
-				'custom_css' => '' !== $base_custom_str
-					? [ 'raw' => base64_encode( $base_custom_str ) ] // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_encode -- Global class custom_css.raw is stored as base64.
-					: null,
-				'null_props' => $base_null_props,
-			];
-		}
-
-		return $variants;
 	}
 
 	private function handle_duplicated_labels( array $intents, array &$touched_items, array $all_labels, Bulk_Operations_Result $results ): void {
@@ -714,101 +625,6 @@ class Manage_Classes_Ability extends Abstract_Ability {
 		$new_order = array_values( $new_order );
 
 		return array_merge( $new_order, $added_ids );
-	}
-
-	private function apply_mode( array $existing, array $new_variants, string $mode, array $affected_breakpoints, array $props_to_remove = [] ): array {
-		if ( 'replace' === $mode ) {
-			$kept      = array_values(
-				array_filter( $existing, fn( $v ) => ! in_array( $v['meta']['breakpoint'] ?? null, $affected_breakpoints, true ) )
-			);
-			$clean_new = array_values(
-				array_filter(
-					array_map(
-						function ( $v ) {
-							unset( $v['null_props'] );
-							return $v;
-						},
-						$new_variants
-					),
-					fn( $v ) => ! empty( $v['props'] ) || ! empty( $v['custom_css'] )
-				)
-			);
-			return array_merge( $kept, $clean_new );
-		}
-
-		$result = $existing;
-
-		foreach ( $new_variants as $new_variant ) {
-			$bp    = $new_variant['meta']['breakpoint'] ?? null;
-			$state = $new_variant['meta']['state'] ?? null;
-			$match = null;
-
-			foreach ( $result as $i => $v ) {
-				if ( ( $v['meta']['breakpoint'] ?? null ) === $bp && ( $v['meta']['state'] ?? null ) === $state ) {
-					$match = $i;
-					break;
-				}
-			}
-
-			if ( null === $match ) {
-				$clean = $new_variant;
-				unset( $clean['null_props'] );
-				if ( ! empty( $clean['props'] ) || ! empty( $clean['custom_css'] ) ) {
-					$result[] = $clean;
-				}
-				continue;
-			}
-
-			$variant_null_props  = $new_variant['null_props'] ?? [];
-			$wipe_all            = in_array( 'all', $variant_null_props, true );
-			$all_props_to_remove = array_merge(
-				$props_to_remove,
-				array_filter( $variant_null_props, fn( $k ) => 'all' !== $k )
-			);
-
-			$merged_props = $wipe_all
-				? ( $new_variant['props'] ?? [] )
-				: array_merge( $result[ $match ]['props'] ?? [], $new_variant['props'] ?? [] );
-
-			foreach ( $all_props_to_remove as $key ) {
-				unset( $merged_props[ $key ] );
-			}
-
-			$merged_props = array_filter( $merged_props, fn( $v ) => null !== $v );
-
-			$merged_custom_css = $wipe_all
-				? ( $new_variant['custom_css'] ?? null )
-				: $this->merge_custom_css( $result[ $match ]['custom_css'] ?? null, $new_variant['custom_css'] ?? null );
-
-			if ( empty( $merged_props ) && null === $merged_custom_css ) {
-				unset( $result[ $match ] );
-				$result = array_values( $result );
-				continue;
-			}
-
-			$result[ $match ]['props']      = $merged_props;
-			$result[ $match ]['custom_css'] = $merged_custom_css;
-			unset( $result[ $match ]['null_props'] );
-		}
-
-		return $result;
-	}
-
-	private function extract_null_reset_props( array $css ): array {
-		return array_keys(
-			array_filter( $css, fn( $v ) => null === $v || 'null' === $v )
-		);
-	}
-
-	private function merge_custom_css( ?array $existing, ?array $incoming ): ?array {
-		$existing_raw = isset( $existing['raw'] ) ? base64_decode( $existing['raw'] ) : ''; // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_decode -- Decoding stored base64 custom_css for merge.
-		$incoming_raw = isset( $incoming['raw'] ) ? base64_decode( $incoming['raw'] ) : ''; // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_decode -- Decoding stored base64 custom_css for merge.
-
-		$merged = trim( $existing_raw . ' ' . $incoming_raw );
-
-		return '' !== $merged
-			? [ 'raw' => base64_encode( $merged ) ] // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_encode -- Storing merged custom_css as base64.
-			: null;
 	}
 
 	protected function build_class_item( string $id, string $label, array $css ) {
