+++ tests/phpunit/elementor/modules/atomic-widgets/css-converter/test-css-converter.php
@@ -181,123 +181,6 @@ class Test_Css_Converter extends TestCase {
 		$this->assertSame( '', $result['customCss'] );
 	}
 
-	public function test_parse_nested__flat_css_returns_single_base_block() {
-		// Arrange.
-		$converter = $this->make( new Converter_Registry() );
-
-		// Act.
-		$result = $converter->parse_nested( 'color: red; font-size: 14px;' );
-
-		// Assert.
-		$this->assertArrayNotHasKey( 'error', $result );
-		$this->assertCount( 1, $result['blocks'] );
-		$this->assertNull( $result['blocks'][0]['selector'] );
-		$this->assertSame( 'color: red; font-size: 14px;', $result['blocks'][0]['css'] );
-	}
-
-	public function test_parse_nested__hover_block_extracted_correctly() {
-		// Arrange.
-		$converter = $this->make( new Converter_Registry() );
-
-		// Act.
-		$result = $converter->parse_nested( 'color: red; &:hover { color: blue; }' );
-
-		// Assert.
-		$this->assertArrayNotHasKey( 'error', $result );
-		$this->assertCount( 2, $result['blocks'] );
-		$this->assertNull( $result['blocks'][0]['selector'] );
-		$this->assertSame( ':hover', $result['blocks'][1]['selector'] );
-		$this->assertSame( ' color: blue; ', $result['blocks'][1]['css'] );
-	}
-
-	public function test_parse_nested__class_selector_block_extracted_correctly() {
-		// Arrange.
-		$converter = $this->make( new Converter_Registry() );
-
-		// Act.
-		$result = $converter->parse_nested( '&.my-class { color: green; }' );
-
-		// Assert.
-		$this->assertArrayNotHasKey( 'error', $result );
-		$this->assertCount( 2, $result['blocks'] );
-		$this->assertSame( '.my-class', $result['blocks'][1]['selector'] );
-		$this->assertSame( ' color: green; ', $result['blocks'][1]['css'] );
-	}
-
-	public function test_parse_nested__multiple_blocks_all_extracted() {
-		// Arrange.
-		$converter = $this->make( new Converter_Registry() );
-		$css = 'color: red; &:hover { color: blue; } &:focus { outline: none; } &:active { color: green; }';
-
-		// Act.
-		$result = $converter->parse_nested( $css );
-
-		// Assert.
-		$this->assertArrayNotHasKey( 'error', $result );
-		$this->assertCount( 4, $result['blocks'] );
-		$this->assertSame( ':hover', $result['blocks'][1]['selector'] );
-		$this->assertSame( ':focus', $result['blocks'][2]['selector'] );
-		$this->assertSame( ':active', $result['blocks'][3]['selector'] );
-	}
-
-	public function test_parse_nested__bare_ampersand_block_returns_error() {
-		// Arrange.
-		$converter = $this->make( new Converter_Registry() );
-
-		// Act.
-		$result = $converter->parse_nested( 'color: red; & { color: blue; }' );
-
-		// Assert.
-		$this->assertSame( [], $result['blocks'] );
-		$this->assertArrayHasKey( 'error', $result );
-		$this->assertNotEmpty( $result['error'] );
-	}
-
-	public function test_parse_nested__unclosed_brace_returns_error() {
-		// Arrange.
-		$converter = $this->make( new Converter_Registry() );
-
-		// Act.
-		$result = $converter->parse_nested( 'color: red; &:hover { color: blue;' );
-
-		// Assert.
-		$this->assertSame( [], $result['blocks'] );
-		$this->assertArrayHasKey( 'error', $result );
-		$this->assertNotEmpty( $result['error'] );
-	}
-
-	public function test_parse_nested__empty_block_content_returned_as_is() {
-		// Arrange.
-		$converter = $this->make( new Converter_Registry() );
-
-		// Act.
-		$result = $converter->parse_nested( '&:hover {}' );
-
-		// Assert.
-		$this->assertArrayNotHasKey( 'error', $result );
-		$this->assertCount( 2, $result['blocks'] );
-		$this->assertSame( ':hover', $result['blocks'][1]['selector'] );
-		$this->assertSame( '', $result['blocks'][1]['css'] );
-	}
-
-	public function test_parse_nested__base_declarations_preserved_after_stripping_blocks() {
-		// Arrange.
-		$converter = $this->make( new Converter_Registry() );
-
-		// Act.
-		$result = $converter->parse_nested( 'color: red; &:hover { color: blue; scroll-behavior: smooth; } font-size: 14px;' );
-
-		// Assert.
-		$this->assertArrayNotHasKey( 'error', $result );
-		$base_css = $result['blocks'][0]['css'];
-		$hover_css = $result['blocks'][1]['css'];
-		$this->assertStringContainsString( 'color: red;', $base_css );
-		$this->assertStringContainsString( 'font-size: 14px;', $base_css );
-		$this->assertStringNotContainsString( '&:hover', $base_css );
-		$this->assertStringContainsString( 'scroll-behavior: smooth;', $hover_css);
-		$this->assertStringContainsString( 'color: blue;', $hover_css);
-	}
-
 	public function test_convert__legit_lookalike_property_is_not_blocked() {
 		// Arrange.
 		$converter = $this->make( new Converter_Registry() );
